more accurate duels

+++ src/Services/PocketWaifu/Fight/AttackInfo.cs
@@ -4,7 +4,7 @@ namespace Sanakan.Services.PocketWaifu.Fight
 {
     public class AttackInfo
     {
-        public int Dmg { get; set; }
+        public double Dmg { get; set; }
         public ulong AtkCardId { get; set; }
         public ulong DefCardId { get; set; }
     }
