+++ src/Services/PocketWaifu/Fight/HpSnapshot.cs
@@ -5,6 +5,6 @@ namespace Sanakan.Services.PocketWaifu.Fight
     public class HpSnapshot
     {
         public ulong CardId { get; set; }
-        public int Hp { get; set; }
+        public double Hp { get; set; }
     }
 }
